添加article接口

+++ src/controllers/user.ts
@@ -2,21 +2,21 @@ import jwt from 'jsonwebtoken';
 import { JWT_SECRET } from '../constants';
 import { Context } from 'koa';
 import * as argon2 from 'argon2';
-import { getManager } from 'typeorm';
+import { getRepository } from 'typeorm';
 import { User } from '../entity/user';
 import { NotFoundException, ForbiddentException, UnauthorizedException } from '../exceptions';
 
 export default class UserController {
   public static async getUsers(ctx: Context) {
     const { query, pagenum, pagesize } = ctx.query;
-    const userRepository = getManager().getRepository(User);
-
+    const userRepository = getRepository(User);
     let [users, total] = await userRepository.findAndCount({
       order: {
         create_time: 'DESC',
       },
       take: pagesize,
-      // skip: 5,
+      skip: (pagenum - 1) * pagesize,
+      // where: { name: query },
     });
 
     ctx.status = 200;
@@ -27,7 +27,7 @@ export default class UserController {
   }
 
   public static async addUser(ctx: Context) {
-    const userRepository = getManager().getRepository(User);
+    const userRepository = getRepository(User);
 
     const newUser = new User();
     newUser.name = ctx.request.body.name;
@@ -51,7 +51,7 @@ export default class UserController {
   }
 
   public static async getUserInfo(ctx: Context) {
-    const userRepository = getManager().getRepository(User);
+    const userRepository = getRepository(User);
 
     const Token = ctx.request.header['authorization'].split(' ')[1];
     let verify_id: any = jwt.verify(Token, JWT_SECRET);
@@ -64,7 +64,7 @@ export default class UserController {
           ...user,
           avatar: 'https://wpimg.wallstcn.com/f778738c-e4f8-4870-b634-56703b4acafe.gif',
           introduction: 'I am a super administrator',
-          roles: ['admin'],
+          roles: ['editor'],
         },
         meta: { message: '获取用户成功', code: 20000 },
       };
@@ -74,7 +74,7 @@ export default class UserController {
   }
 
   public static async showUserDetail(ctx: Context) {
-    const userRepository = getManager().getRepository(User);
+    const userRepository = getRepository(User);
     const user = await userRepository.findOne(+ctx.params.id);
 
     if (user) {
@@ -96,7 +96,7 @@ export default class UserController {
     //   throw new ForbiddentException();
     // }
 
-    const userRepository = getManager().getRepository(User);
+    const userRepository = getRepository(User);
 
     await userRepository.update(id, { email, mobile, update_time: new Date() });
     const updatedUser = await userRepository.findOne(+ctx.request.body.id);
@@ -110,18 +110,20 @@ export default class UserController {
   }
 
   public static async deleteUser(ctx: Context) {
-    console.log('ctx:', ctx.params);
-
     const userId = +ctx.params.id;
     console.log(userId, +ctx.state.user.id);
     // if (userId !== +ctx.state.user.id) {
     //   throw new ForbiddentException();
     // }
 
-    const userRepository = getManager().getRepository(User);
-    await userRepository.delete(+ctx.params.id);
-
-    ctx.status = 200;
-    ctx.body = { meta: { message: '删除成功', code: 20000 } };
+    const userRepository = getRepository(User);
+    const user = await userRepository.findOne(+ctx.params.id);
+    if (User.name == 'admin') {
+      throw new NotFoundException('不能删除admin账号');
+    } else {
+      await userRepository.delete(+ctx.params.id);
+      ctx.status = 200;
+      ctx.body = { meta: { message: '删除成功', code: 20000 } };
+    }
   }
 }
